+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 1.0.3 - TBD
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- Nothing.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- CS fixes (package handler)
+- Localize $@ variable (package handler)
+- Couldn't create symlink for the Roundcube webmail (package handler)
+
 ## 1.0.2 - 20190410
 
 ### Added